test(flow): distribute edge-rs compatibility coverage
Move the compatibility assertions into the relevant existing flow suites and remove the standalone compatibility test file.
