+++ tests/flow/test_slowlog.py
@@ -1,4 +1,5 @@
 import asyncio
+import re
 from common import *
 from falkordb.asyncio import FalkorDB
 from distutils.version import StrictVersion
@@ -48,6 +49,7 @@ class testSlowLog():
         # slow log should contain a single entry, no duplicates
         slowlog = self.graph.slowlog()
         self.env.assertEquals(len(slowlog), 1)
+        self.env.assertTrue(re.fullmatch(r"\d+", str(slowlog[0][0])) is not None)
 
         # saturate slowlog
         self.populate_slowlog(20)
@@ -271,4 +273,3 @@ class testSlowLog():
         queries = [entry[2] for entry in entries]
         self.env.assertIn (q0, queries)
         self.env.assertIn (q1, queries)
-
